+++ CHANGELOG.md
@@ -5,7 +5,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.2.0](https://github.com/TissUUmaps/TissUUmaps-schema/compare/v1.1.0...v1.2.0)
+
+New schema versions: 1.2, 0.2
+
+### Added
+- New Project.menuButtons field (v0, v1)
+- New Filter enum items: SplitChannel, Colormap (v1)
+- New Layer.clip field (v1)
+- New Project.collectionLayout field (v1)
+- LayerFilter values can be strings, booleans or numbers (v0, v1)
+- Allow list values: ExpectedHeader.cb_gr_list, ExpectedHeader.pie_list, ExpectedHeader.shape_gr_list (v1)
+- Fix default uid value for ExpectedHeader (v0)
+
+### Fixed
+
+### Changed
+- Allow extra fields in v0
+- Allow dictionary values: ExpectedHeader.cb_gr_dict, ExpectedHeader.pie_dict, ExpectedHeader.shape_gr_dict (v1)
+
+### Removed
+
+
 ## [v1.1.0](https://github.com/TissUUmaps/TissUUmaps-schema/compare/v1.0.2...v1.1.0)
+
+New schema versions: 1.1
 
 ### Added
 - New field `ExpectedHeader.stroke_width`
@@ -20,15 +44,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [v1.0.2](https://github.com/TissUUmaps/TissUUmaps-schema/compare/v1.0.1...v1.0.2)
 
-### Changed
-- Renamed tissuumaps_schema.utils.current_schema_module to tissuumaps_schema.utils.CURRENT_SCHEMA_MODULE
-
 ### Added
 - Exported tissuumaps_schema.utils.CURRENT_SCHEMA_MODULE as tissuumaps_schema.current
 - Exported some additional functionality (e.g. get_major_version) via tissuumaps_schema.utils
 
+### Fixed
+
+### Changed
+- Renamed tissuumaps_schema.utils.current_schema_module to tissuumaps_schema.utils.CURRENT_SCHEMA_MODULE
+
+### Removed
+
 
 ## [v1.0.1](https://github.com/TissUUmaps/TissUUmaps-schema/compare/v1.0.0...v1.0.1)
+
+### Added
 
 ### Fixed
 - Bug in `models` command
@@ -36,8 +66,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Versioning scheme: PATCH indicates Python package bugfixes (removed PATCH from schema versioning)
 
+### Removed
+
 
 ## [v1.0.0](https://github.com/TissUUmaps/TissUUmaps-schema/compare/v0.1.0...v1.0.0)
+
+New schema versions: 1.0, 0.1
 
 ### Added
 - Schema version 0.1.0 ("expectedCSV" format of an EOL TissUUmaps version)